restructuring make_dataset fixing early stopping adding get data for test set
