+++ src/data/make_dataset.py
@@ -6,7 +6,7 @@ from pathlib import Path
 from torch.utils.data import Dataset, DataLoader
 import logging
 from torchvision import transforms
-
+import sys
 import torch
 from PIL import Image
 
@@ -41,30 +41,66 @@ class Lego_Dataset(torch.utils.data.Dataset):
 
 
 
-def main():
-    """ Runs data processing scripts to turn raw data from (../raw) into
-        cleaned data ready to be analyzed (saved in ../processed).
-    """
-    #ToDo Paths
+def get_data():
 
-    parent = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-
-    data_path = os.path.join(parent,'data/external/lego_dataset/')
+    current_script_directory = os.path.dirname(os.path.abspath(__file__))
+    parent_directory = os.path.abspath(os.path.join(current_script_directory, '..'))
+    root_directory = os.path.abspath(os.path.join(parent_directory, '..'))
+    sys.path.append(parent_directory)
 
 
-    index = pd.read_csv(os.path.join(data_path,Path('index.csv')))
-    labels = index["class_id"]-1
-    files = index["path"]
+    
+    data_path = os.path.join(root_directory,"data", "external", "lego_dataset")
 
-    transform = transforms.Compose([transforms.ToTensor(),
-                                    transforms.Normalize((0.5,), (0.5,))])
+    index = pd.read_csv(os.path.join(data_path, 'index.csv'))
 
 
+    #Train Validation Split
+    train_index = index.sample(int(0.75*len(index.index)))
 
-    trainset = Lego_Dataset(file_paths=files, path = data_path, labels=labels,transform=transform)
+    remaining_indices = list(set(index.index) - set(train_index.index))
+    # Create a new DataFrame with the remaining indices
+    train_index.reset_index(inplace= True, drop=True)
+    val_index = index.loc[remaining_indices].reset_index(drop=True)
 
-    torch.save(trainset,os.path.join(parent,Path('data/processed/trainset.pth')))
-    #train_loader = DataLoader(trainset, batch_size=32, shuffle=True)
+    train_labels = train_index["class_id"]-1
+    train_files = train_index["path"]
 
-    logger = logging.getLogger(__name__)
-    logger.info('making final data set from raw data')
+    val_labels = val_index["class_id"]-1
+    val_files = val_index["path"]
+
+
+    train_transforms = transforms.Compose([
+    transforms.RandomResizedCrop(224),  # Random crop and resize to 224x224
+    transforms.RandomHorizontalFlip(),  # Randomly flip the image horizontally
+    transforms.RandomRotation(degrees=30),  # Randomly rotate the image up to 30 degrees
+    transforms.RandomAffine(degrees=0, translate=(0.1, 0.1)),  # Randomly translate the image
+    transforms.RandomPerspective(distortion_scale=0.5, p=0.5),  # Apply perspective transformation
+    transforms.ToTensor(),  # Convert the image to a PyTorch tensor
+    transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]),  # ImageNet statistics
+    ])
+
+    # Define transforms for validation and test data (typically no augmentation)
+    val_transforms = transforms.Compose([
+        transforms.Resize(256),  # Resize to 256x256
+        transforms.CenterCrop(224),  # Center crop to 224x224
+        transforms.ToTensor(),  # Convert the image to a PyTorch tensor
+        transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]),  # ImageNet statistics
+    ])
+
+
+    return Lego_Dataset(file_paths=train_files, path = data_path, labels=train_labels,transform=train_transforms),Lego_Dataset(file_paths=val_files, path = data_path, labels=val_labels,transform=val_transforms)
+
+
+def get_test_Data():
+    current_script_directory = os.path.dirname(os.path.abspath(__file__))
+    parent_directory = os.path.abspath(os.path.join(current_script_directory, '..'))
+    root_directory = os.path.abspath(os.path.join(parent_directory, '..'))
+    sys.path.append(parent_directory)
+
+
+    
+    data_path = os.path.join(root_directory,"data", "external", "lego_dataset")
+
+    index = pd.read_csv(os.path.join(data_path, 'test.csv'))
+
