+++ src/models/train_model.py
@@ -9,6 +9,12 @@ import hydra
 import torchvision.transforms as transforms
 from torch.utils.data import DataLoader
 import wandb
+from src.data.make_dataset import get_data
+
+
+current_script_directory = os.path.dirname(os.path.abspath(__file__))
+parent_directory = os.path.abspath(os.path.join(current_script_directory, '..'))
+root_directory = os.path.abspath(os.path.join(parent_directory, '..'))
 
 wandb.init(
     # set the wandb project where this run will be logged
@@ -23,56 +29,6 @@ wandb.init(
     }
 )
 
-# Import LegoDataset class (automated relative import)
-current_script_directory = os.path.dirname(os.path.abspath(__file__))
-parent_directory = os.path.abspath(os.path.join(current_script_directory, '..'))
-root_directory = os.path.abspath(os.path.join(parent_directory, '..'))
-sys.path.append(parent_directory)
-
-from data.make_dataset import Lego_Dataset
-
-
-def get_data():
-    data_path = os.path.join(root_directory,"data", "external", "lego_dataset")
-
-    index = pd.read_csv(os.path.join(data_path, 'index.csv'))
-
-
-    #Train Validation Split
-    train_index = index.sample(int(0.75*len(index.index)))
-
-    remaining_indices = list(set(index.index) - set(train_index.index))
-    # Create a new DataFrame with the remaining indices
-    train_index.reset_index(inplace= True, drop=True)
-    val_index = index.loc[remaining_indices].reset_index(drop=True)
-
-    train_labels = train_index["class_id"]-1
-    train_files = train_index["path"]
-
-    val_labels = val_index["class_id"]-1
-    val_files = val_index["path"]
-
-
-    train_transforms = transforms.Compose([
-    transforms.RandomResizedCrop(224),  # Random crop and resize to 224x224
-    transforms.RandomHorizontalFlip(),  # Randomly flip the image horizontally
-    transforms.RandomRotation(degrees=30),  # Randomly rotate the image up to 30 degrees
-    transforms.RandomAffine(degrees=0, translate=(0.1, 0.1)),  # Randomly translate the image
-    transforms.RandomPerspective(distortion_scale=0.5, p=0.5),  # Apply perspective transformation
-    transforms.ToTensor(),  # Convert the image to a PyTorch tensor
-    transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]),  # ImageNet statistics
-])
-
-    # Define transforms for validation and test data (typically no augmentation)
-    val_transforms = transforms.Compose([
-        transforms.Resize(256),  # Resize to 256x256
-        transforms.CenterCrop(224),  # Center crop to 224x224
-        transforms.ToTensor(),  # Convert the image to a PyTorch tensor
-        transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]),  # ImageNet statistics
-    ])
-
-
-    return Lego_Dataset(file_paths=train_files, path = data_path, labels=train_labels,transform=train_transforms),Lego_Dataset(file_paths=val_files, path = data_path, labels=val_labels,transform=val_transforms)
 
 @hydra.main(config_name="model_config.yaml")
 def train_model(cfg):
@@ -91,7 +47,7 @@ def train_model(cfg):
 
 
     best_val_loss = float('inf')
-    patience = 2  # Number of epochs to wait for improvement
+    patience = 5  # Number of epochs to wait for improvement
     epochs_without_improvement = 0
 
     print("Training start...")
@@ -150,8 +106,8 @@ def train_model(cfg):
 
         wandb.log({'train_acc' : epoch_accuracy,'val_acc' : val_epoch_accuracy,'train_loss' : batch_loss,'val_loss' : val_loss})
 
-        if val_loss < best_val_loss:
-            best_val_loss = val_loss
+        if val_epoch_loss < best_val_loss:
+            best_val_loss = val_epoch_loss
             epochs_without_improvement = 0
             
             #save the best model checkpoint here
