+++ wandb/run-20230929_160922-2f2t1o5f/files/wandb-summary.json
@@ -1,0 +1,1 @@
+{"train_acc": 0.1423611111111111, "val_acc": 0.3958333333333333, "train_loss": 2.5526342391967773, "val_loss": 2.0708024501800537, "_timestamp": 1695996584.1273682, "_runtime": 21.26625418663025, "_step": 0, "_wandb": {"runtime": 20}}
