+++ wandb/run-20230929_161049-6zx0y4rv/files/wandb-summary.json
@@ -1,0 +1,1 @@
+{"train_acc": 0.8263888888888888, "val_acc": 0.875, "train_loss": 0.8586603403091431, "val_loss": 0.6522802710533142, "_timestamp": 1695996769.0114398, "_runtime": 119.53609585762024, "_step": 5, "_wandb": {"runtime": 118}}
