+++ wandb/run-20230929_161423-vrx0vx4t/files/config.yaml
@@ -1,0 +1,47 @@
+wandb_version: 1
+
+learning_rate:
+  desc: null
+  value: 0.003
+architecture:
+  desc: null
+  value: mobilenetv3
+dataset:
+  desc: null
+  value: Lego-Minifigures
+epochs:
+  desc: null
+  value: 10
+_wandb:
+  desc: null
+  value:
+    python_version: 3.9.18
+    cli_version: 0.15.11
+    framework: torch
+    is_jupyter_run: false
+    is_kaggle_kernel: false
+    start_time: 1695996863.94566
+    t:
+      1:
+      - 1
+      - 41
+      - 49
+      - 50
+      - 55
+      - 63
+      2:
+      - 1
+      - 41
+      - 49
+      - 50
+      - 55
+      - 63
+      3:
+      - 16
+      - 23
+      4: 3.9.18
+      5: 0.15.11
+      8:
+      - 3
+      - 5
+      13: windows-amd64
