+++ wandb/run-20230929_161423-vrx0vx4t/files/requirements.txt
@@ -1,0 +1,95 @@
+alabaster==0.7.13
+antlr4-python3-runtime==4.9.3
+appdirs==1.4.4
+asttokens==2.4.0
+babel==2.12.1
+backcall==0.2.0
+backports.functools-lru-cache==1.6.5
+certifi==2023.7.22
+charset-normalizer==3.2.0
+click==8.1.7
+colorama==0.4.6
+comm==0.1.4
+coverage==7.3.1
+debugpy==1.6.7
+decorator==5.1.1
+docker-pycreds==0.4.0
+docutils==0.20.1
+exceptiongroup==1.1.3
+executing==1.2.0
+filelock==3.12.4
+flake8==6.1.0
+fsspec==2023.9.2
+gitdb==4.0.10
+gitpython==3.1.37
+huggingface-hub==0.17.3
+hydra-core==1.3.2
+idna==3.4
+imagesize==1.4.1
+importlib-metadata==6.8.0
+ipykernel==6.25.2
+ipython==8.15.0
+jedi==0.19.0
+jinja2==3.1.2
+jupyter-client==8.3.1
+jupyter-core==5.3.2
+markupsafe==2.1.3
+matplotlib-inline==0.1.6
+mccabe==0.7.0
+mpmath==1.3.0
+nest-asyncio==1.5.6
+networkx==3.1
+numpy==1.26.0
+omegaconf==2.3.0
+packaging==23.1
+pandas==2.1.1
+parso==0.8.3
+pathtools==0.1.2
+pickleshare==0.7.5
+pillow==10.0.1
+pip==23.2.1
+platformdirs==3.10.0
+prompt-toolkit==3.0.39
+protobuf==4.24.3
+psutil==5.9.0
+pure-eval==0.2.2
+pycodestyle==2.11.0
+pyflakes==3.1.0
+pygments==2.16.1
+python-dateutil==2.8.2
+python-dotenv==1.0.0
+pytz==2023.3.post1
+pywin32==305.1
+pyyaml==6.0.1
+pyzmq==23.2.1
+requests==2.31.0
+safetensors==0.3.3
+sentry-sdk==1.31.0
+setproctitle==1.3.2
+setuptools==68.2.2
+six==1.16.0
+smmap==5.0.1
+snowballstemmer==2.2.0
+sphinx==7.2.6
+sphinxcontrib-applehelp==1.0.7
+sphinxcontrib-devhelp==1.0.5
+sphinxcontrib-htmlhelp==2.0.4
+sphinxcontrib-jsmath==1.0.1
+sphinxcontrib-qthelp==1.0.6
+sphinxcontrib-serializinghtml==1.1.9
+src==0.1.0
+stack-data==0.6.2
+sympy==1.12
+timm==0.9.7
+torch==2.0.1
+torchvision==0.15.2
+tornado==6.2
+tqdm==4.66.1
+traitlets==5.10.1
+typing-extensions==4.8.0
+tzdata==2023.3
+urllib3==2.0.5
+wandb==0.15.11
+wcwidth==0.2.6
+wheel==0.41.2
+zipp==3.17.0
