+++ wandb/run-20230929_161423-vrx0vx4t/files/wandb-summary.json
@@ -1,0 +1,1 @@
+{"train_acc": 0.8819444444444444, "val_acc": 0.8888888888888888, "train_loss": 0.3672159016132355, "val_loss": 0.47770145535469055, "_timestamp": 1695997114.4955032, "_runtime": 250.54984307289124, "_step": 12, "_wandb": {"runtime": 249}}
